+++ grunt-tasks/options/karma.js
@@ -11,7 +11,6 @@ var files = [
     'demo/bower_components/lodash/dist/lodash.compat.js',
     'demo/bower_components/moment/moment.js',
     'demo/bower_components/html2canvas/build/html2canvas.js',
-    'demo/bower_components/mousetrap/mousetrap.js',
     'demo/bower_components/angular-hotkeys/build/hotkeys.min.js',
     'demo/bower_components/ng-debounce/angular-debounce.js',
     'demo/bower_components/angular-mocks/angular-mocks.js',