+++ src/components/rxSearchBox/scripts/rxSearchBox.page.js
@@ -147,27 +147,9 @@ exports.rxSearchBox = {
      * @description Page object representing the rxSearchBox object.
      */
     initialize: function (rxSearchBoxElement) {
-        if (rxSearchBoxElement === undefined) {
-            rxSearchBoxElement = $('rx-search-box');
-        }
-
         rxSearchBox.rootElement = {
             get: function () { return rxSearchBoxElement; }
         };
         return Page.create(rxSearchBox);
-    },
-
-    /**
-     * @memberof rxSearchBox
-     * @deprecated Use {@link rxSearchBox.initialize} without arguments instead.
-     * @type {rxSearchBox}
-     * @description Will return a page object representing the _first_ rxSearchBox object found on the page.
-     */
-    main: (function () {
-        rxSearchBox.rootElement = {
-            get: function () { return $$('rx-search-box').first(); }
-        };
-        return Page.create(rxSearchBox);
-    })()
-
+    }
 };